Cambio a BigDecimal

+++ src/main/java/dds/monedero/model/Movimiento.java
@@ -1,21 +1,20 @@
 package dds.monedero.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Movimiento {
   private LocalDate fecha;
-  //En ningún lenguaje de programación usen jamás doubles para modelar dinero en el mundo real
-  //siempre usen numeros de precision arbitraria, como BigDecimal en Java y similares
-  private double monto; // TODO: ya que se hace refactor se debe cambiar a BigDecimal
+  private BigDecimal monto;
   private boolean esDeposito;
 
-  public Movimiento(LocalDate fecha, double monto, boolean esDeposito) {
+  public Movimiento(LocalDate fecha, BigDecimal monto, boolean esDeposito) {
     this.fecha = fecha;
     this.monto = monto;
     this.esDeposito = esDeposito;
   }
 
-  public double getMonto() {
+  public BigDecimal getMonto() {
     return monto;
   }
